deploy/requirements.txt
fastapi==0.108.0
uvicorn==0.27.0
python-dotenv==1.0.0
supabase==1.2.0
httpx==0.24.1 